update tests and examples
