+++ internal/sdkv2/morpheus/resources/script/boot_script_resource_test.go
@@ -3,13 +3,65 @@
 package script_test
 
 import (
+	"context"
+	"os"
 	"testing"
 
-	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
-
+	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
+	"github.com/hashicorp/terraform-plugin-mux/tf5to6server"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
+	sdkv2morpheus "github.com/HPE/terraform-provider-hpe/internal/sdkv2/morpheus"
 )
+
+// RenderBootScriptConfig renders a Terraform configuration for boot_script resource.
+// It accepts a name and a map of overrides to customize the default field values.
+func RenderBootScriptConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":    name,
+		"Content": "ls",
+	}
+
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	return testhelpers.RenderExample(
+		t,
+		"morpheus_boot_script_resource.tf.tmpl",
+		"Name", defaults["Name"],
+		"Content", defaults["Content"],
+	)
+}
+
+func TestMain(m *testing.M) {
+	code := m.Run()
+
+	testhelpers.WriteMergedResults()
+
+	os.Exit(code)
+}
+
+func newProviderWithError() (tfprotov6.ProviderServer, error) {
+	return tf5to6server.UpgradeServer(
+		context.Background(),
+		sdkv2morpheus.Provider().GRPCProvider,
+	)
+}
+
+var testAccProtoV6ProviderFactories = map[string]func() (
+	tfprotov6.ProviderServer, error,
+){
+	"hpe": newProviderWithError,
+}
 
 func TestAccMorpheusBootScriptExampleOk(t *testing.T) {
 	t.Parallel()
@@ -24,10 +76,9 @@ func TestAccMorpheusBootScriptExampleOk(t *testing.T) {
 
 	name := acctest.RandomWithPrefix(t.Name())
 
-	resourceConfig, err := testhelpers.RenderExample(t, "morpheus_boot_script_resource.tf.tmpl",
-		"Name", name,
-		"Content", "ls",
-	)
+	resourceConfig, err := RenderBootScriptConfig(t, name, map[string]string{
+		"Content": "ls",
+	})
 	if err != nil {
 		t.Fatal(err)
 	}
