+++ internal/sdkv2/morpheus/resources/script/morpheus_preseed_script_resource_test.go
@@ -3,26 +3,26 @@
 package script_test
 
 import (
-	"context"
-	"os"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
-	"github.com/hashicorp/terraform-plugin-mux/tf5to6server"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 
 	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
-	sdkv2morpheus "github.com/HPE/terraform-provider-hpe/internal/sdkv2/morpheus"
 )
 
-// RenderMorpheusPreseedScriptConfig renders a Terraform configuration for morpheus_preseed_script resource.
-// It accepts a map of overrides to customize the default field values.
-func RenderMorpheusPreseedScriptConfig(t *testing.T, overrides map[string]string) (string, error) {
+// RenderMorpheusPreseedScriptConfig renders a Terraform configuration
+// for morpheus_preseed_script resource.
+// It accepts a name and a map of overrides to customize the default field values.
+func RenderMorpheusPreseedScriptConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
 	t.Helper()
 
 	defaults := map[string]string{
-		"Name":    acctest.RandomWithPrefix(t.Name()),
+		"Name":    name,
 		"Content": "ls",
 	}
 
@@ -38,24 +38,6 @@ func RenderMorpheusPreseedScriptConfig(t *testing.T, overrides map[string]string
 	)
 }
 
-func TestMain(m *testing.M) {
-	code := m.Run()
-
-	testhelpers.WriteMergedResults()
-
-	os.Exit(code)
-}
-
-func newProviderWithError() (tfprotov6.ProviderServer, error) {
-	return tf5to6server.UpgradeServer(context.Background(), sdkv2morpheus.Provider().GRPCProvider)
-}
-
-var testAccProtoV6ProviderFactories = map[string]func() (
-	tfprotov6.ProviderServer, error,
-){
-	"hpe": newProviderWithError,
-}
-
 func TestAccMorpheusPreseedScriptExampleOk(t *testing.T) {
 	t.Parallel()
 
@@ -69,8 +51,7 @@ func TestAccMorpheusPreseedScriptExampleOk(t *testing.T) {
 
 	name := acctest.RandomWithPrefix(t.Name())
 
-	resourceConfig, err := RenderMorpheusPreseedScriptConfig(t, map[string]string{
-		"Name":    name,
+	resourceConfig, err := RenderMorpheusPreseedScriptConfig(t, name, map[string]string{
 		"Content": "ls",
 	})
 	if err != nil {
